# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSKit Research Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/mcp/tools": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "summary": "List Mcp Tools",
        "description": "List discovered tools allowed for the authenticated user.",
        "operationId": "list_mcp_tools_api_v1_mcp_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpTool"
                  },
                  "type": "array",
                  "title": "Response List Mcp Tools Api V1 Mcp Tools Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mcp/tools/{name}/invoke": {
      "post": {
        "tags": [
          "capabilities"
        ],
        "summary": "Invoke Mcp Tool",
        "description": "Validate and invoke an allowed MCP tool with optional idempotency.",
        "operationId": "invoke_mcp_tool_api_v1_mcp_tools__name__invoke_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Arguments"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpInvokeResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-runs": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "summary": "List Tool Runs",
        "description": "List the user's saved MCP tool invocation records.",
        "operationId": "list_tool_runs_api_v1_tool_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ToolRun"
                  },
                  "type": "array",
                  "title": "Response List Tool Runs Api V1 Tool Runs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-runs/{run_id}/project": {
      "patch": {
        "tags": [
          "capabilities"
        ],
        "summary": "Save Tool Run To Project",
        "description": "Associate an owned tool invocation with an accessible project.",
        "operationId": "save_tool_run_to_project_api_v1_tool_runs__run_id__project_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRunMove"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/af3/jobs": {
      "post": {
        "tags": [
          "capabilities"
        ],
        "summary": "Submit Af3 Job",
        "description": "Reserve GPU quota and submit a member's AF3 job.",
        "operationId": "submit_af3_job_api_v1_af3_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Af3JobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Af3Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/af3/jobs/{job_id}": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "summary": "Get Af3 Job",
        "description": "Read an AF3 job belonging to the authenticated user.",
        "operationId": "get_af3_job_api_v1_af3_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Af3Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "capabilities"
        ],
        "summary": "Cancel Af3 Job",
        "description": "Cancel an owned AF3 job and its associated agent run.",
        "operationId": "cancel_af3_job_api_v1_af3_jobs__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Af3Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Af3FoldInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "modelSeeds": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Modelseeds"
          },
          "sequences": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 32,
            "minItems": 1,
            "title": "Sequences"
          },
          "dialect": {
            "type": "string",
            "const": "alphafold3",
            "title": "Dialect"
          },
          "version": {
            "type": "integer",
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Version"
          },
          "bondedAtomPairs": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bondedatompairs"
          },
          "userCCD": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userccd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "modelSeeds",
          "sequences",
          "dialect",
          "version"
        ],
        "title": "Af3FoldInput"
      },
...
      "Af3Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "estimated_gpu_minutes": {
            "type": "integer",
            "title": "Estimated Gpu Minutes"
          },
          "actual_gpu_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Gpu Minutes"
          },
          "gpu_accounting_status": {
            "type": "string",
            "enum": [
              "reserved",
              "released",
              "settled",
              "pending_reconciliation",
              "reconciled"
            ],
            "title": "Gpu Accounting Status",
            "default": "reserved"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "artifacts": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "fold_input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fold Input"
          },
          "simulation": {
            "type": "boolean",
            "title": "Simulation",
            "default": false
          },
          "resource_requirements": {
            "$ref": "#/components/schemas/ComputeResourceRequirements"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "progress",
          "estimated_gpu_minutes"
        ],
        "title": "Af3Job"
      },
      "Af3JobRequest": {
        "properties": {
          "estimated_gpu_minutes": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Estimated Gpu Minutes",
            "default": 20
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "fold_input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Af3FoldInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "Af3JobRequest"
      },
...
      "ComputeResourceRequirements": {
        "properties": {
          "capability": {
            "type": "string",
            "const": "af3",
            "title": "Capability",
            "default": "af3"
          },
          "gpu_count": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Gpu Count",
            "default": 1
          },
          "min_gpu_memory_mb": {
            "type": "integer",
            "maximum": 1048576.0,
            "minimum": 0.0,
            "title": "Min Gpu Memory Mb",
            "default": 0
          }
        },
        "type": "object",
        "title": "ComputeResourceRequirements"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "McpInvokeResult": {
        "properties": {
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "status": {
            "type": "string",
            "const": "completed",
            "title": "Status",
            "default": "completed"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "result"
        ],
        "title": "McpInvokeResult"
      },
      "McpTool": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "input_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Schema"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description"
        ],
        "title": "McpTool"
      },
...
      "ToolRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tool",
          "title",
          "arguments",
          "result",
          "created_at"
        ],
        "title": "ToolRun"
      },
      "ToolRunMove": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "ToolRunMove"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}